Added aptfile
This file is optional while deploying to heroku and is only needed if the application to be deployed uses opencv. It is required since opencv has some additional dependencies that are not installed on heroku by default and needs to be installed manually. These dependencies can be installed by using linux commands sudo apt get install or sudo apt install. So an additional Aptfile with the names of those dependencies helps with this and tells heroku to install these dependencies using the apt buildpack that was added while creating app.